Add MedSAM2 latest to model dict (#19)
MedSAM2 provides its own checkpoints and configs. The config files have
been copied directly (from bowang-lab/MedSAM2@332f30d), the checkpoints
linked to.

A bit crude for the moment, if there are likely to be other extension
models to be included then we'll have to make things more neat and
futureproof